widen int multiplication to long to prevent overflow (NarrowCalculation)

+++ controller/src/main/java/com/linbit/linstor/core/apicallhandler/controller/CtrlQuerySizeInfoHelper.java
@@ -343,7 +343,7 @@ public class CtrlQuerySizeInfoHelper
             if (cacheEntry != null)
             {
                 long now = System.currentTimeMillis();
-                if (cacheEntry.cacheTimestampInMs + ignoreCacheOlderThanSecRef * SEC_TO_MS > now)
+                if (cacheEntry.cacheTimestampInMs + (long) ignoreCacheOlderThanSecRef * SEC_TO_MS > now)
                 {
                     ret = new PairNonNull<>(cacheEntry.obj, (now - cacheEntry.cacheTimestampInMs) * 1.0 / SEC_TO_MS);
                 }
